requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
google-generativeai
alembic
authlib
PyPDF2
python-multipart